10.18 set record to 128 bytes

+++ Project2/include/bpt.h
@@ -127,12 +127,11 @@ int height(node * root);
 int path_to_root(node * root, node * child);
 void print_leaves(node * root);
 void print_tree(node * root);
-void find_and_print(node * root, int key, bool verbose); 
-void find_and_print_range(node * root, int range1, int range2, bool verbose); 
-int find_range(node * root, int key_start, int key_end, bool verbose,
-		int returned_keys[], void * returned_pointers[]); 
-node * find_leaf(node * root, int key, bool verbose);
-record * find(node * root, int key, bool verbose);
+void find_and_print(node * root, long long int key, bool verbose); 
+void find_and_print_range(node * root, long long int range1, long long int range2, bool verbose); 
+int find_range(node * root, long long int key_start, long long int key_end, bool verbose, int returned_keys[], void * returned_pointers[]); 
+node * find_leaf(node * root, long long int key, bool verbose);
+record * find(node * root, long long int key, bool verbose);
 int cut(int length);
 
 // Insertion.
@@ -141,18 +140,17 @@ record * make_record(int value);
 node * make_node(void);
 node * make_leaf(void);
 int get_left_index(node * parent, node * left);
-node * insert_into_leaf( node * leaf, int key, record * pointer );
-node * insert_into_leaf_after_splitting(node * root, node * leaf, int key,
+node * insert_into_leaf( node * leaf, long long int key, record * pointer );
+node * insert_into_leaf_after_splitting(node * root, node * leaf, long long int key,
 										record * pointer);
 node * insert_into_node(node * root, node * parent, 
-		int left_index, int key, node * right);
+		int left_index, long long int key, node * right);
 node * insert_into_node_after_splitting(node * root, node * parent,
-										int left_index,
-		int key, node * right);
-node * insert_into_parent(node * root, node * left, int key, node * right);
-node * insert_into_new_root(node * left, int key, node * right);
-node * start_new_tree(int key, record * pointer);
-node * insert( node * root, int key, int value );
+										int left_index, long long int key, node * right);
+node * insert_into_parent(node * root, node * left, long long int key, node * right);
+node * insert_into_new_root(node * left, long long int key, node * right);
+node * start_new_tree(long long int key, record * pointer);
+node * insert( node * root, long long int key, int value );
 
 // Deletion.
 
@@ -163,8 +161,8 @@ node * coalesce_nodes(node * root, node * n, node * neighbor,
 node * redistribute_nodes(node * root, node * n, node * neighbor,
 						  int neighbor_index,
 		int k_prime_index, int k_prime);
-node * delete_entry(node * root, node * n, int key, void * pointer);
-node * delete(node * root, int key);
+node * delete_entry(node * root, node * n, long long int key, void * pointer);
+node * delete(node * root, long long int key);
 
 void destroy_tree_nodes(node * root);
 node * destroy_tree(node * root);
